+++ src/Services/Kros.ToDos.Api/Application/Commands/DeleteToDo/DeleteToDoCommand.cs
@@ -15,7 +15,7 @@ namespace Kros.ToDos.Api.Application.Commands
         /// <param name="id">ToDo id.</param>
         /// <param name="userId">User id.</param>
         /// <param name="organizationId">Organization id.</param>
-        public DeleteToDoCommand(int id, int userId, int organizationId)
+        public DeleteToDoCommand(long id, long userId, long organizationId)
         {
             Id = id;
             UserId = userId;
@@ -25,18 +25,18 @@ namespace Kros.ToDos.Api.Application.Commands
         /// <summary>
         /// Id.
         /// </summary>
-        public int Id { get; }
+        public long Id { get; }
 
         /// <summary>
         /// User Id.
         /// </summary>
         [JsonIgnore]
-        public int UserId { get; }
+        public long UserId { get; }
 
         /// <summary>
         /// Organization id.
         /// </summary>
         [JsonIgnore]
-        public int OrganizationId { get; }
+        public long OrganizationId { get; }
     }
 }
