+++ src/Services/Kros.ToDos.Api/Application/Queries/PipeLines/IUserResourceQuery.cs
@@ -8,11 +8,11 @@
         /// <summary>
         /// User id.
         /// </summary>
-        int UserId { get; }
+        long UserId { get; }
 
         /// <summary>
         /// Organization id.
         /// </summary>
-        int OrganizationId { get; }
+        long OrganizationId { get; }
     }
 }
